- FIELD_OFFSET returns LONG, not LONG_PTR or size_t. - Define TYPE_ALIGNMENT for msvc. - Define DbgPrintEx the same way in NDK and DDK - C_ASSERT is a typedef, so don't put it inside functions as msvc doesn't allow this. - Kernel now *compiles* with RosBE + msvc _and_ our DDK. Linker errors from the ninth circle of hell remains.
svn path=/trunk/; revision=42216

+++ reactos/include/host/typedefs.h
@@ -219,9 +219,9 @@ typedef const UNICODE_STRING *PCUNICODE_STRING;
 
 #define NT_SUCCESS(x)           ((x)>=0)
 #if !defined(__GNUC__)
-#define FIELD_OFFSET(t,f)       ((LONG_PTR)&(((t*)0)->f))
+#define FIELD_OFFSET(t,f)       ((LONG)(LONG_PTR)&(((t*) 0)->f))
 #else
-#define FIELD_OFFSET(t,f)       __builtin_offsetof(t,f)
+#define FIELD_OFFSET(t,f)       ((LONG)__builtin_offsetof(t,f))
 #endif
 #define RTL_CONSTANT_STRING(s)  { sizeof(s)-sizeof((s)[0]), sizeof(s), s }
 #define CONTAINING_RECORD(address, type, field)  ((type *)(((ULONG_PTR)address) - (ULONG_PTR)(&(((type *)0)->field))))
